Implemented 'cannot receive a gift more than once in a round' logic. The Player Card composable now calls stealGift() function, which returns an error if the gift was already received during the current round.
If an error is returned instead of a successful steal, an alert dialog displays the error to the player.

This required setting up a list of giftsReceivedThisRound to track the gifts a player claims per round.Additionally updated the onStartNewRound() function to clear the list of all gifts accept the currently claimed gift.